+++ processor/src/fast/sys_ops.rs
@@ -95,9 +95,17 @@ impl FastProcessor {
         if let Some(system_event) = SystemEvent::from_event_id(event_id) {
             handle_system_event(process, system_event, err_ctx)
         } else {
-            host.on_event(process, event_id)
+            let clk = process.clk();
+            let mutations = host
+                .on_event(&*process, event_id)
                 .await
-                .map_err(|err| ExecutionError::event_error(err, event_id, err_ctx))
+                .map_err(|err| ExecutionError::event_error(err, event_id, err_ctx))?;
+            for mutation in mutations {
+                self.advice
+                    .apply_mutation(mutation)
+                    .map_err(|err| ExecutionError::advice_error(err, clk, err_ctx))?;
+            }
+            Ok(())
         }
     }
 }